Filters `find`, `find_index`, `has` and `reject`
