+++ liquid2/builtin/__init__.py
@@ -43,25 +43,19 @@ from .expressions import parse_positional_and_keyword_arguments
 from .expressions import parse_primitive
 from .expressions import parse_string_or_identifier
 from .expressions import parse_string_or_path
-from .filters.array import compact  # noqa: F401
 from .filters.array import concat
 from .filters.array import first
 from .filters.array import join
 from .filters.array import last
-from .filters.array import map_  # noqa: F401
 from .filters.array import reverse
-from .filters.array import sort  # noqa: F401
-from .filters.array import sort_natural  # noqa: F401
-from .filters.array import sort_numeric  # noqa: F401
-from .filters.array import sum_  # noqa: F401
-from .filters.array import uniq  # noqa: F401
-from .filters.array import where  # noqa: F401
 from .filters.babel import Currency
 from .filters.babel import DateTime
 from .filters.babel import Number
 from .filters.babel import Unit
-from .filters.compact_arrow import CompactFilter
-from .filters.map_arrow import MapFilter
+from .filters.filtering_filters import CompactFilter
+from .filters.filtering_filters import RejectFilter
+from .filters.filtering_filters import WhereFilter
+from .filters.map_filter import MapFilter
 from .filters.math import abs_
 from .filters.math import at_least
 from .filters.math import at_most
@@ -77,9 +71,9 @@ from .filters.misc import JSON
 from .filters.misc import date
 from .filters.misc import default
 from .filters.misc import size
-from .filters.sort_arrow import SortFilter
-from .filters.sort_arrow import SortNaturalFilter
-from .filters.sort_arrow import SortNumericFilter
+from .filters.sorting_filters import SortFilter
+from .filters.sorting_filters import SortNaturalFilter
+from .filters.sorting_filters import SortNumericFilter
 from .filters.string import append
 from .filters.string import capitalize
 from .filters.string import downcase
@@ -106,15 +100,14 @@ from .filters.string import truncatewords
 from .filters.string import upcase
 from .filters.string import url_decode
 from .filters.string import url_encode
-from .filters.sum_arrow import SumFilter
+from .filters.sum_filter import SumFilter
 from .filters.translate import BaseTranslateFilter
 from .filters.translate import GetText
 from .filters.translate import NGetText
 from .filters.translate import NPGetText
 from .filters.translate import PGetText
 from .filters.translate import Translate
-from .filters.uniq_arrow import UniqFilter
-from .filters.where_arrow import WhereFilter
+from .filters.uniq_filter import UniqFilter
 from .loaders.caching_file_system_loader import CachingFileSystemLoader
 from .loaders.choice_loader import CachingChoiceLoader
 from .loaders.choice_loader import ChoiceLoader
@@ -259,22 +252,15 @@ def register_default_tags_and_filters(env: Environment) -> None:  # noqa: PLR091
     env.filters["first"] = first
     env.filters["last"] = last
     env.filters["concat"] = concat
-    # env.filters["map"] = map_
     env.filters["map"] = MapFilter()
     env.filters["reverse"] = reverse
-    # env.filters["sort"] = sort
     env.filters["sort"] = SortFilter()
-    # env.filters["sort_natural"] = sort_natural
     env.filters["sort_natural"] = SortNaturalFilter()
-    # env.filters["sort_numeric"] = sort_numeric
     env.filters["sort_numeric"] = SortNumericFilter()
-    # env.filters["sum"] = sum_
     env.filters["sum"] = SumFilter()
-    # env.filters["where"] = where
     env.filters["where"] = WhereFilter()
-    # env.filters["uniq"] = uniq
+    env.filters["reject"] = RejectFilter()
     env.filters["uniq"] = UniqFilter()
-    # env.filters["compact"] = compact
     env.filters["compact"] = CompactFilter()
 
     env.filters["abs"] = abs_
